Refactor Autofill script to use SourceProvider
