+++ Tests/BrowserServicesKitTests/Autofill/AutofillEmailUserScriptTests.swift
@@ -38,9 +38,9 @@ class AutofillEmailUserScriptTests: XCTestCase {
         """.data(using: .utf8)!
         let privacyConfig = AutofillTestHelper.preparePrivacyConfig(embeddedConfig: embeddedConfig)
         let properties = ContentScopeProperties(gpcEnabled: false, sessionKey: "1234")
-        
-        return AutofillUserScript(privacyConfigurationManager: privacyConfig, properties: properties,
-                                  encrypter: MockEncrypter(), hostProvider: SecurityOriginHostProvider())
+        let sourceProvider = DefaultAutofillSourceProvider(privacyConfigurationManager: privacyConfig,
+                                                           properties: properties)
+        return AutofillUserScript(scriptSourceProvider: sourceProvider, encrypter: MockEncrypter(), hostProvider: SecurityOriginHostProvider())
     }()
     let userContentController = WKUserContentController()
